# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "description": "Cette API permet de gérer les différentes composantes de la course d'orientation fixe de Bex.",
    "version": "1.0.0",
    "title": "API Course orientation Bex"
  },
  "servers": [
    {
      "url": "https://cobex-uu6t.onrender.com/api",
      "description": "URL online, elle peut mettre jusqu'à une minute pour répondre"
    },
    {
      "url": "http://localhost:3000/api",
      "description": "URL local pour faire des tests"
    }
  ],
...
  "paths": {
...
    "/resultats": {
      "post": {
        "summary": "Créer un nouveau résultat",
        "tags": [
          "resultats"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ResultatData"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Résultat créé avec succès",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Resultat"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/NotConnected"
          },
          "500": {
            "description": "Résultat invalide",
            "content": {
              "text/html": {
                "example": "Resultat validation failed: trailID: Path `trailID` is required."
              }
            }
          }
        }
      },
      "get": {
        "summary": "Récupérer les résultats",
        "tags": [
          "resultats"
        ],
        "parameters": [
          {
            "name": "utilisateurs",
            "in": "query",
            "description": "Filtrer par ID utilisateur",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "parcours",
            "in": "query",
            "description": "Filtrer par ID parcours (uniquement s'il y a un utilisateur défini)",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Liste des utilisateurs récupérée avec succès",
            "content": {
              "application/json": {
                "examples": {
                  "withoutInclude": {
                    "summary": "Sans paramètres de requête",
                    "value": 
                      [
                        {
                          "id": "678504c0224b7e662ab9620a",
                          "temps": 120
                        },
                        {
                          "id": "678504c0224b7e662ab9620b",
                          "temps": 130
                        },{
                          "id": "678504c0224b7e662ab9620c",
                          "temps": 200
                      }
                      ]
                  },
                  "withInclude": {
                    "summary": "Avec paramètres `utilisateurs=678504c0224b7e662ab961e6&parcours=678504c0224b7e662ab961ff`",
                    "value": 
                    [
                      {
                          "id": "678504c0224b7e662ab9620c",
                          "temps": 200,
                          "parcours": {
                              "id": "678504c0224b7e662ab961ff",
                              "nom": "Parcours moyen"
                          }
                      }
                  ]
                  }
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "responses": {
...
      "NotConnected": {
        "description": "Personne non connectée",
        "content": {
          "text/html": {
            "example": "Authorization header is missing"
          }
        }
      },
...
    },
    "schemas": {
...
      "Resultat": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "Identifiant unique du résultat",
            "example": "6cf8e6fedc4cf8b5ace5a689"
          },
          "temps": {
            "type": "integer",
            "description": "Temps en secondes mis pour faire le parcours",
            "minimum": 0,
            "exclusiveMinimum": true,
            "example": "600"
          },
          "userID": {
            "type": "string",
            "description": "L'identifiant unique de la personne qui a fait ce temps",
            "example": "7038e6fedc4cf8b5ace5a689"
          },
          "trailID": {
            "type": "string",
            "description": "L'identifiant unique du parcours finalisé",
            "example": "00f8e6fedc4cf8b5ace5a689"
          },
          "createdAt": {
            "type": "string",
            "description": "Date de création du résultat",
            "format": "date-time",
            "example": "2022-01-01T00:00:00Z"
          }
        },
        "required": [
          "id",
          "temps",
          "userID",
          "trailID",
          "createdAt"
        ]
      },
      "ResultatData": {
        "properties": {
          "temps": {
            "type": "integer",
            "description": "Temps en secondes mis pour faire le parcours",
            "minimum": 0,
            "exclusiveMinimum": true,
            "example": "600"
          },
          "trailID": {
            "type": "string",
            "description": "L'identifiant unique du parcours finalisé",
            "example": "00f8e6fedc4cf8b5ace5a689"
          }
        },
        "required": [
          "temps",
          "trailID"
        ]
      },
...
    }
  }
}